Modules.json file restructured for easier C# parsing
The file structure was too complex for the C# client and would have required too much maintenance and caused confusion. It was OK for JavaScript, but now with the client in C#, the file should be simplified. It is now OK for modules to have the same name, as long as their versions are different.